kafka-bigquery/requirements.txt
confluent-kafka
google-cloud-bigquery
avro-python3
radar
Faker